# CDev::poll() performance improvements