--- a/src/modules/uORB/uORB_tests/CMakeLists.txt
+++ b/src/modules/uORB/uORB_tests/CMakeLists.txt
@@ -44,6 +44,7 @@ px4_add_module(
 	MODULE modules__uORB__uORB_tests
 	MAIN uorb_tests
 	STACK_MAIN 2048
+	PRIORITY "SCHED_PRIORITY_MAX"
 	COMPILE_FLAGS
 	SRCS ${SRCS}
 	DEPENDS