--- a/src/modules/uORB/uORB_tests/uORBTest_UnitTest.hpp
+++ b/src/modules/uORB/uORB_tests/uORBTest_UnitTest.hpp
@@ -146,8 +146,8 @@ int uORBTest::UnitTest::latency_test(orb_id_t T, bool print)
 	// prevent access if the caller data goes out of scope
 	int pubsub_task = px4_task_spawn_cmd("uorb_latency",
 					     SCHED_DEFAULT,
-					     SCHED_PRIORITY_MAX - 5,
-					     1500,
+					     SCHED_PRIORITY_MAX,
+					     1700,
 					     (px4_main_t)&uORBTest::UnitTest::pubsubtest_threadEntry,
 					     args);