--- a/src/systemcmds/top/CMakeLists.txt
+++ b/src/systemcmds/top/CMakeLists.txt
@@ -34,7 +34,7 @@ px4_add_module(
 	MODULE systemcmds__top
 	MAIN top
 	STACK_MAIN 1700
-	PRIORITY 255
+	PRIORITY "SCHED_PRIORITY_MAX"
 	COMPILE_FLAGS
 	SRCS
 		top.c