When looking at the uorb latency test with `uorb_tests latency_test` the latency for publish-poll-wakeup is 40 us. This seems small, but running at 1kHz it accounts for 4% CPU load for a single pub-sub pair.

This PR reduces the latency to 30us by avoiding the use of an expensive semaphore and disabling IRQs instead (the actual critical section is very small).
It also improves the `uorb_tests latency_test` for better output and more consistent results.

master:
```
uORB note: ---------------- LATENCY TEST ------------------
INFO  [uorb_tests] mean:     40.4320 us
INFO  [uorb_tests] std dev:   1.3466 us
INFO  [uorb_tests] min:      39 us
INFO  [uorb_tests] max:      57 us
INFO  [uorb_tests] missed topic updates: 0
```
This PR:
```
uORB note: ---------------- LATENCY TEST ------------------
INFO  [uorb_tests] mean:     30.4120 us
INFO  [uorb_tests] std dev:   1.0637 us
INFO  [uorb_tests] min:      29 us
INFO  [uorb_tests] max:      45 us
INFO  [uorb_tests] missed topic updates: 0
```

The changes here will become relevant when increasing the rate controller loop to 1kHz. I quickly tested running the attitude controller and fmu at 1kHz and the difference in CPU load is 4% (35 vs 31).

I've been using and flying with these changes for over a week.

Changes are split into separate commits with further explanations.

### Jitter

One interesting side note is looking at the latency after stopping mavlink (`mavlink stop-all`):
```
uORB note: ---------------- LATENCY TEST ------------------
INFO  [uorb_tests] mean:     30.1050 us
INFO  [uorb_tests] std dev:   0.3288 us
INFO  [uorb_tests] min:      29 us
INFO  [uorb_tests] max:      34 us
INFO  [uorb_tests] missed topic updates: 0
```
It can be seen that there is quite a bit less jitter that comes from UART interrupts/critical sections. I think having UART TX DMA would help here.