@dagar 2018-07-25
Looks good.

Just a thought, but if we really wanted to optimize orb polling I think we could get away with a single semaphore per cdev and avoid the bulk of poll waiter setup and teardown (at least on nuttx).