@dagar 2018-07-25
> It can be seen that there is quite a bit less jitter that comes from UART interrupts/critical sections. I think having UART TX DMA would help here.

@davids5 FYI