@davids5 2018-07-25
@bkueng 
If this is just a uOrb - should it be moved there and done differently? 
Is the dynamic allocation warranted?
px4_enter/exit_critical section is a macros and does not need #ifdeffing.
the disable and re enable around the malloc may not be needed. What are we trying to do? Avoid a context switch? Would schedule lock be a better choice?