@bkueng 2018-07-25
> @bkueng out of curiosity did you fly at 1kHz? If so could you share the log?

I obviously had to test it :) But you don't want to see the code. It just the bare minimum required for acro and in-flight switching between 250 and 1000 Hz. A log is here: https://logs.px4.io/plot_app?log=74d50cdc-3ee6-4534-bbfe-2e9576514323
As for the flight behavior: to my surprise it does make a small difference. You'll notice it in situations where you have strong prop-wash effects for instance (the vehicle is more stable). I only flew LOS, I can imagine that the effect is stronger noticeable in FPV.
My goal is to implement it properly anytime soon, and after that we'll need the >=4kHz IMU sampling.

----

@davids5 The goal here is to maximize performance and minimize memory overhead. Which is why there's a dynamic allocation (that is rarely used). I was thinking about a linked list but there is nothing where we could add the next pointer (without changing NuttX).
malloc uses a semaphore, which is why the IRQs are re-enabled for that.